// src/components/forms/checkins/monthly/types.ts
export type MonthlyFormValues = {
  currentlyFilledPositions: string;
  currentVacancies: string;
  resignationsThisMonth: string;
  newHiresAndAcademyGraduates: string;
  averageTimeToFillDays: string;
  leadershipLevelVacancies: string;
  totalOvertimeHours: string;
  averageShiftLengthHours: string;
  shiftCoverageShortages: string;
  mandatoryOvertimePercentage: string;
  plannedOvertimePercentage: string;
  unplannedOvertimePercentage: string;
  callInAndHoldoverIncidents: string;
  totalSickLeaveDaysUsed: string;
  employeesOnFmlaLeave: string;
  newFmlaRequests: string;
  workersCompClaimsFiled: string;
  peerSupportActivations: string;
  criticalIncidentExposures: string;
  lineOfDutyDeathsOrSeriousInjuries: string;
  lineOfDutyDeathsOrSeriousInjuriesCount: string;
  leadershipMoraleRating: string;
  frontlineMoraleRating: string;
  disciplinaryActions: string;
  formalGrievancesFiled: string;
  promotionsOrLeadershipDevelopmentCount: string;
  dataConfidence: string;
  overtimeBudgetUtilizationPercentage: string;
  hiringBudgetAvailability: string;
  staffingBudgetConstraint: string;
  totalCallsOrIncidents: string;
  responseTimeStandardsMet: string;
  specialtyUnitVacancies: string;
  minimumRestPeriodRequirementMet: string;
  ptoVacationAccrualBacklog: string;
  returnToDutyIncidentsBeforeFullRecovery: string;
  eapReferralsOrUtilizations: string;
  topLeadershipConcern: string;
  topLeadershipConcernOther: string;
  additionalContextOrNotes: string;
};